remove dead code from simulator
