+++ Simulator.py
@@ -17,6 +17,7 @@ from Ball import Ball
 from LinearAlegebraUtils import distBetween
 from RunAtBallBrain import RunAtBallBrain
 from Team import Team
+from matplotlib import animation
 
 
 #Called once for initialization
@@ -35,23 +36,10 @@ class Simulator(object):
         self.currWP = 0
         self.ballWPs = [array([50.0, -100.0, 0.0]), array([0.0, 100.0, -70.0]), array([50.0, 20.0, 100.0]),array([-30.0, 50.0, -100.0]), array([80.0, -50.0, 50.0]), array([80.0, -50.0, -50.0]), array([-65.0, 20.0, 50.0]), array([-50.0, 20.0, -60.0])]
         self.fig = plt.figure(figsize=(16,12))
-        self.timeStep = 1/float(30)
-        self.frameProb = float(self.fps) / 30
-        self.currTime = float(0)
-        self.drawIndex = 0
-        self.physicsIndex = 0
 
 
-    def setup(self):    
-        #setup directory to save the images
-        self.imageDirName = 'images'
-        try:
-            os.mkdir(self.imageDirName)
-        except:
-            print self.imageDirName + " subdirectory already exists. OK."
-
-  
-         #define teams which the agents can be a part of
+    def setup(self):
+        #define teams which the agents can be a part of
         teamA = Team("A", '#ff99ff')
         teamB = Team("B", '#ffcc99')
         #Defining a couple of agents 
@@ -131,49 +119,20 @@ class Simulator(object):
 
         return self.loop(ax), plt.gca().set_ylim(ax.get_ylim()[::-1])
 
-    def run(self):
-        #Run setup once
-        self.setup()
+def tick(i):
+    return sim.tick()
 
-        while(self.currTime < self.simTime):
-            self.fixedLoop()
-            currProb = float(self.drawIndex)/float(self.physicsIndex+1)
-            if currProb < self.frameProb:
-                self.drawFrame(self.drawIndex)
-                self.drawIndex+=1
-            self.physicsIndex+=1
-            self.currTime+=float(self.timeStep)
-     
-        print "Physics ran for "+str(self.physicsIndex)+" steps"
-        print "Drawing ran for "+str(self.drawIndex)+" steps"
-            
-    def drawFrame(self, loopIndex):
-        self.fig.clf()
-        ax = self.fig.add_subplot(111, projection='3d')
-        ax.view_init(elev = 30)
-        ax.set_xlabel("X")
-        ax.set_ylabel("Y")
-        ax.set_zlabel("Z")
-        fname = self.imageDirName + "/img{:08d}.png".format(int(loopIndex)) # name the file
-        self.loop(ax)
-        plt.gca().set_ylim(ax.get_ylim()[::-1])
-        self.fig.savefig(fname, format='png', bbox_inches='tight')
-        print 'Written Frame No.'+ str(loopIndex)+' to '+ fname
+FPS = 30 # hardcoded into simulation assumptions
 
+world = World(100, 100)
+sim = Simulator(world, 10, FPS, "images")
 
+# simulator assumes its physics simulation is running at dt = 1/30 s
+# would have to recalculate the true animation framerate (to get physically accurate speeds, etc)
+# if you change the raw sample rate of the animator function (which accepts arguments in ms)
+simlen = 10 # sec
+frames = FPS * simlen
 
-#Simulation runs here
-#set the size of the world
-# world = World(100, 100)
-#specify which world to simulate, total simulation time, and framerate for video
-# sim = Simulator(world, 60, 30, "images")
-#run the simulation
-# sim.run()
+ani = animation.FuncAnimation(sim.fig, tick, frames=frames, blit=True, init_func=sim.setup)
 
-'''
-To create a video using the image sequence, execute the following command in command line.
->ffmpeg -f image2 -i "1%08d.png" -r 30 outPut.mp4
-Make sure to set your current working directory to /images and have ffmpeg in your path.
-'''
-
-    
+ani.save('output.mp4', fps=FPS, extra_args=['-vcodec', 'libx264'])
